Improve silencing of findimports
findimports is a noisy library that prints a bunch of gunk while it's
building the dependency graph. We were replace sys.stdout/stderr with
BytesIO, which was causing an error to be thrown deep inside
findimports. This changes things to use a StringIO instead and to retry
loading the dependency graph if an exception is thrown during the first
try rather than silently failing.