Fix error when fetching error context from files
If there wasn't enough context either before or after an error returned
by mypy (e.g. if the error was near the beginning or end of the file) we would
throw an error when trying to format the context string. This makes the
context gathering easier to read and more robust. It also provides an
option for clients of MypyTask to disable error context gathering if
it's not relevant to them.